backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
pymongo==4.6.1
Pillow==10.2.0
stegano==0.9.9
